backendflights/requirements.txt
amqp==5.0.9
asgiref==3.4.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0
django-filter==21.1
djangorestframework==3.13.1
kombu==5.2.3
prompt-toolkit==3.0.24
pycodestyle==2.8.0
pytz==2021.3
redis==2.10.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
vine==5.0.0
wcwidth==0.2.5
